Maths

SymPh.Maths — Module

Maths

A module containing symbolic representations of differential forms, and operations on them, and a way to convert them to array operations

Abstract Objects

These are abstract object representing differential forms and vectors

SymPh.Maths.Form — Type

Form{D,P<:Primality}

Arguments

D::Integer : degree (0/1/2... form)

P::Primality : The primality of the form

SymPh.Maths.Vect — Type

Vect{P<:Primality}

A vector field

Variables

The variables upon which we act with our operators. Typically our prognostic variables will be defined like this since they are not defined as an operation on something.

SymPh.Maths.FormVariable — Type

FormVariable{D,P}(name::String, save::Bool) <: Form{D,P}

A variable representing a named differential form

SymPh.Maths.VectorVariable — Type

VectorVariable{P}(name::String, save::Bool) <: Vect{P}

A variable representing a named vector field

Operators

These are the operators acting on our variables. Their type is the type of the result of the operation.

The Wedge product $\wedge$

SymPh.Maths.Wedge — Type

Wedge{Dl + Dr,P}(name::String, save::Bool, left::Form{Dl,P}, right::Form{Dr,P}) <: Form{Dl + Dr, P}

Wedge product of two forms TODO Wedge between different primalities ?

The Exterior Derivative $\mathrm{d}$

SymPh.Maths.ExteriorDerivative — Type

ExteriorDerivative{D,P}(name::String, save::Bool, omega::Form{D-1,P}) <: Form{D,P}

The exterior derivative $\mathrm{d}\omega$

Transforms a $k$-form into a $k+1$-form

The exterior derivative $\mathrm{d}\omega$ of a $k$-form $\omega$ is a $k+1$-form.

In 2D,

\[\mathrm{d}F^0 = \partial_x f \mathrm{d} x+\partial_y f \mathrm{d} y\]

Corresponds to the gradient operation on a scalar field.

\[\mathrm{d}U^1 = (\partial_x u_2-\partial_y u_1) \mathrm{d} x \wedge \mathrm{d} y\]

Corresponds to the curl of a vector field.

\[\mathrm{d} W^2 = 0\]

The Codifferential $\delta$

SymPh.Maths.Codifferential — Type

Codifferential{D,P}(name::String, save::Bool, omega::Form{D+1,P}) <: Form{D,P}

The codifferential $\delta \omega$

Transforms a $k$-form into a $k-1$-form

The Interior Product with a vector field $\iota_\mathbf{X}$

SymPh.Maths.InteriorProduct — Type

InteriorProduct{D, Pv, Pf}(name::String, save::Bool, X::Vect, omega::Form, interp = Nothing) <: Form{D, Pf}

The contraction of a $k$-form $\omega$ with a vector field $\mathbf{X}$ which gives us a $k-1$-form $\iota_\mathbf{X}\omega$

Possibility to specify interpolation function

Let $\mathbf{X}$ be a vector field. Then :

\[\iota_\mathbf{X} : \Omega^k (M) \mapsto \Omega^{k-1} (M)\]

We have $\iota_\mathbf{X} (\alpha)= \alpha (\mathbf{X})$, and $\iota_\mathbf{X} ( \alpha \wedge \beta) = ( \iota_\mathbf{X} \alpha) \wedge \beta + (-1)^{|\alpha|} \alpha \wedge ( \iota_\mathbf{X} \beta )$ with $|\alpha|=k$ the degree of the $k$-form $\alpha$.

Let $\mathbf{X}=x_1 \mathbf{i} + x_2 \mathbf{j}$, then (in 2D),

\[\iota_\mathbf{X} F^0 = 0\]

\[\iota_\mathbf{X} U^1 = \iota_\mathbf{X}(u_1 \mathrm{d} x+u_2 \mathrm{d} y)=u_1 \iota_\mathbf{X} \mathrm{d} x+u_2 \iota_\mathbf{X} \mathrm{d} y = u_1x_1+u_2 x_2\]

Corresponds to scalar product of two vector fields.

\[\iota_\mathbf{X} W^2 = \iota_\mathbf{X}(w \mathrm{d} x \wedge \mathrm{d} y) = w( \iota_\mathbf{X}( \mathrm{d} x) \wedge \mathrm{d} y - \mathrm{d} x \wedge \iota_\mathbf{X}( \mathrm{d} y)) = w(x_1 \mathrm{d} y-x_2 \mathrm{d} x)\]

Corresponds to a kind of "perpendicular product" of a scalar field with a vector field.

The Sharp $\sharp$

SymPh.Maths.Sharp — Type

Sharp{P}(name::String, save::Bool, form::Form{1, P}) <: Vect{P}

Corresponds to an application of the metric

The Hodge star $\star$

SymPh.Maths.Hodge — Type

Hodge{D, P}(name::String, save::Bool, form::Form) <: Form{D, P}

Brings a $k$-form to a $n-k$ form and goes from dual to primal and inversely

The inverse of a Laplacian

SymPh.Maths.InverseLaplacian — Type

InverseLaplacian<D, P>(name::String, save::Bool, form::Form{D, P}) <: Form{D, P}

Represents the solution to a Poisson problem

Others

SymPh.Maths.Addition — Type

Addition{D,P}(name::String, save::Bool, left::Form{D,P}, right::Form{D,P}) <: Form{D,P}

The element-wise sum $left + right$.

SymPh.Maths.Substraction — Type

Substraction{D,P}(name::String, save::Bool, left::Form{D,P}, right::Form{D,P}) <: Form{D,P}

The substraction $left - right$.

SymPh.Maths.Negative — Type

Negative{D,P}(name::String, save::Bool, form::Form{D,P}) <: Form{D,P}

The inverse of a form.

SymPh.Maths.FuncCall — Type

FuncCall{D, P}(name::String, save::Bool, func, args::Vector{Form}) <: Form{D,P}

Represents a call to func applied to the objects represented by args. Forces the computation of its arguments.

SymPh.Maths.Division — Type

Division{D,P}(name::String, save::Bool, left::Form{D,P}, right::Form) <: Form{D,P}

TODO What is that actually in terms of FORMS ? Simple division by a scalar field proxied by a 0-form ?

SymPh.Maths.RealProdForm — Type

RealProdForm{D, P}(name::String, save::Bool, real::Real, form::Form{D, P}) <: Form{D, P}

real time form

From Forms to Arrays

SymPh.Maths.to_kernel — Function

to_kernel(exprs...; explparams = ExplicitParam(), verbose=false, bcs = [])

Takes one or multiple expressions and returns a kernel that computes them. The kernel has signature compute!(mesh::Mesh, state::State).

Keyword Arguments

explparams::ExplicitParam : An object holding the parameters like which interpolation to use, etc... (see ExplicitParam)

verbose::Bool : Wether or not to print the functions being generated, the dependency tree, etc...

bcs : A list of objects representing our boundary conditions (WIP)

SymPh.Maths.explicit — Function

explicit(expr::Expression, param = ExplicitParam())

Returns a SymPh.Arrays Expression representing the operation expr on an array, pulling numerical methods from param

SymPh.Maths.ExplicitParam — Type

ExplicitParam(;interp = Arrays.upwind, fvtofd = Arrays.fvtofd2, fdtofv = Arrays.fdtofv2)

The parameters used when necessary to choose numerical methods