Miscellaneous

SymPh.State — Type

State(mesh)

Initializes a State object that holds a list of arrays corresponding to the characteristics of mesh. Whenever trying to access a field, the state checks if the object is in its fields, in which case it returns it, and otherwise it allocates it and then returns it.

SymPh.reset_state — Function

reset_state(state::State)

Deletes all the fields in state

Model

SymPh.Model — Type

Model(rhs!, mesh, state, prognostics; integratorstep! = rk3step!, cfl = 0.6, dtmax = 1, Umax = nothing)

Represents an equation along with a mesh, its saved state and an integrator

SymPh.step! — Function

step!(model::Model; n=1, tend=-1)

Performs n integration steps of the model, and stops at tend if it is a strictly positive value

SymPh.run! — Function

run!(model; ...)

TODO document

SymPh.profile_model! — Function

profile_model!(model)

Runs PProf to profile model

Time integrators

SymPh.euler_forwardstep! — Function

euler_forwardstep!(dt, mesh, state, progs)

Arguments

dt : time increment

mesh : Mesh object

state : State object

progs : Name of prognostic variables

SymPh.rk3step! — Function

rk3step!(dt, mesh, state, progs)

Arguments

dt : time increment

mesh : Mesh object

state : State object

progs : Name of prognostic variables

SymPh.rk4step! — Function

rk4step!(dt, mesh, state, progs)

Arguments

dt : time increment

mesh : Mesh object

state : State object

progs : Name of prognostic variables

Other

SymPh.@Let — Macro

@Let name = foo

Evaluates the expression foo and sets its name to name, then assigns a local variable named appropriately

SymPh.plotform — Function

plotform(form, mesh, state)

Plots form using the metric from mesh and the data from state