Arrays

SymPh.Arrays — Module

Arrays

A module containing symbolic representations of arrays, and operations on them, and a way to convert them to computational kernels

Objects and Variables

Abstract types

SymPh.Arrays.Expression — Type

Expression

Generic Expression

SymPh.Arrays.Atom — Type

Atom <: Expression

An atom is a singular element holding a value (Variable or not)

SymPh.Arrays.Literal — Type

Literal <: Atom

Literal constant value

SymPh.Arrays.Variable — Type

Variable <: Atom

Any variable value

Concrete types

SymPh.Arrays.RealValue — Type

RealValue(name::String, val::Real) <: Literal

A constant real value

SymPh.Arrays.ScalarVariable — Type

ScalarVariable(name::String) <: Variable

A variable holding a single scalar value

SymPh.Arrays.ArrayVariable — Type

ArrayVariable(name::String, depi::Integer, depj::Integer) <: Variable

Array object representing a variable name and a relative position

Operations on arrays

SymPh.Arrays.Operator — Type

Operator <: Expression

Generic operation on an expression

SymPh.Arrays.FuncCall — Type

FuncCall(name::String, func, args::Vector{Expression}, depi::Integer, depj::Integer) <: Operator

Forces computation of its arguments, then calls a function on them For now the function must have signature f(out, args...) where out will be replaced by the name assigned to the operator

SymPh.Arrays.UnaryOperator — Type

UnaryOperator

Any operator of a single argument

SymPh.Arrays.Negative — Type

Negative

Represents the negation of an expression

SymPh.Arrays.AbsoluteValue — Type

AbsoluteValue

SymPh.Arrays.BinaryOperator — Type

BinaryOperator

Any operator on two expressions

SymPh.Arrays.Addition — Type

Addition

SymPh.Arrays.Substraction — Type

Substraction

SymPh.Arrays.Multiplication — Type

Multiplication

SymPh.Arrays.Division — Type

Division

TODO error handling

SymPh.Arrays.Exponentiation — Type

Exponentiation

SymPh.Arrays.BooleanExpression — Type

BooleanExpression

SymPh.Arrays.UnaryBooleanOperator — Type

UnaryBooleanOperator

SymPh.Arrays.BinaryBooleanOperator — Type

BinaryBooleanOperator

SymPh.Arrays.TernaryOperator — Type

TernaryOperator

Symbolic representation of a ? b : c

SymPh.Arrays.GreaterThan — Type

GreaterThan

tests if left > right

Meshes

SymPh.Arrays.Mesh — Type

Mesh(ni, nj, nh, mgr, msk, xc, yc, xv, yv; iperio=false, jperio=false)

TODO document

SymPh.Arrays.CartesianMesh — Function

CartesianMesh(ni, nj, nh, mgr, Lx = 1, Ly = 1; xperio=false, yperio=false)

A rectangular mesh of extent Lx, Ly

SymPh.Arrays.PolarMesh — Function

PolarMesh(ni, nj, nh, mgr, msk, rin = 0.5, rout = 1.5)

An annulus mesh with inner radius rin and outer radius rout

The i/x direction is the radial, j/y orthoradial

Interpolations

SymPh.Arrays.upwind — Function

upwind(U::Expression, a::Expression, o::Expression, lr::String, dir::String)

Five point upwind interpolation

Arguments

U : The transportant velocity

a : The object to upwind

o : The order of interpolation at that point

lr : If lr=="left", compute a[i-0.5], else a[i+0.5]

dir : along i or j

SymPh.Arrays.avg2pt — Function

avg2pt(U::Expression, a::Expression, o::Expression, lr::String, dir::String)

Two point average interpolation

Arguments

U : The transportant velocity, unused

a : The object to interpolate

o : The order of interpolation at that point, unused

lr : If lr=="left", compute a[i-0.5], else a[i+0.5]

dir : along i or j

SymPh.Arrays.weno — Function

weno(U::Expression, a::Expression, o::Expression, lr::String, dir::String)

Weno interpolation of fifth order

Arguments

U : The transportant velocity

a : The object to upwind

o : The order of interpolation at that point

lr : If lr=="left", compute a[i-0.5], else a[i+0.5]

dir : along i or j

Finite volumes and finite differences

SymPh.Arrays.fvtofd2 — Function

fvtofd2(q, msk, dir)

A second order function to take a finite volume quantity q to finite difference in the direction dir, using msk to detect borders

SymPh.Arrays.fvtofd4 — Function

fvtofd4(q, msk, dir)

A fourth order function to take a finite volume quantity q to finite difference in the direction dir, using msk to detect borders

SymPh.Arrays.fdtofv2 — Function

fdtofv2(q, msk, dir)

A second order function to take a finite difference quantity q to finite volume in the direction dir, using msk to detect borders

SymPh.Arrays.fdtofv4 — Function

fdtofv4(q, msk, dir)

A fourth order function to take a finite difference quantity q to finite volume in the direction dir, using msk to detect borders

From arrays to kernels

SymPh.Arrays.DepNode — Type

DepNode(name::String, expr::Union{Nothing, Expression})

A node of a dependency tree of expressions made of expressions

SymPh.Arrays.Block — Type

Block

Any operation or block of operations that can be done at once

SymPh.Arrays.LoopBlock — Type

LoopBlock

A block of operations to be computed in a loop on our domain

SymPh.Arrays.CallBlock — Type

CallBlock

A block representing a single function call

SymPh.Arrays.Sequence — Type

Sequence

A sequence of blocks of operations

SymPh.Arrays.to_kernel — Function

to_kernel(seq::Sequence, fill; verbose = false)

Converts seq into a computing kernel, fill is the list of names of value that need their halo filled.